fix: receive ux cleanup - message for sender, layout, spacing, qr size
